+++ Changelog.md
@@ -1,4 +1,5 @@
 # Changelog for mona-lista
 
 ## Unreleased changes
+- (b.2) Implemented ``swapColumn``.
 - (b.1) Implemented ``swapRow``.